pyproject.toml
[project]
name = "rag-context-relevance-evaluator"
version = "0.0.0"
description = "ML3 RAG Context Relevance Evaluator"
authors = [
    {name = "MLcube", email = "info@mlcube.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "guardrails-ai>=0.4.0"
]

[project.optional-dependencies]
dev = [
    "arize-phoenix>=7.12.0",
    "datasets>=3.2.0",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "pyright>=1.1.393",
    "pytest>=8.3.4",
    "ruff>=0.9.5",
    "scikit-learn>=1.6.1",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]
